+++ tests/test_validate.py
@@ -91,7 +91,7 @@ def test_url_repr():
     assert (
         repr(validate.URL(relative=False, error=None)) ==
         '<URL(relative=False, error={0!r})>'
-        .format('"{input}" is not a valid URL.')
+        .format('Invalid URL.')
     )
     assert (
         repr(validate.URL(relative=True, error='foo')) ==
@@ -110,6 +110,8 @@ def test_url_repr():
     "!#$%&'*+-/=?^_`{}|~@example.org",
     'niceandsimple@[64.233.160.0]',
     'niceandsimple@localhost',
+    u'josé@blah.com',
+    u'δοκ.ιμή@παράδειγμα.δοκιμή',
 ])
 def test_email_valid(valid_email):
     validator = validate.Email()
@@ -145,7 +147,7 @@ def test_email_repr():
     assert (
         repr(validate.Email(error=None)) ==
         '<Email(error={0!r})>'
-        .format('"{input}" is not a valid email address.')
+        .format('Invalid email address.')
     )
     assert (
         repr(validate.Email(error='foo')) ==